Add previously released API to jacoco-stubs-gen
This is needed because Metalava is going to be strict about requiring a
previously released API to ensure that flagged APIs work correctly.
Module `jacoco-stubs-gen` does not use flagged APIs or provide
compatibility guarantees on its API but it is also using `droidstubs`
in a very different way to most other uses in the build and it is not
worth special casing it.

Flag: EXEMPT build change to improve flag support
Bug: 425309457
Test: TH
Change-Id: I9c74c231ad5d07e2543c109130a9b1830b11ed46